requirements.txt
fastapi
uvicorn
torch
torchvision
albumentations
opencv-python
pillow
rasterio